controller code factored out and working
